+++ src/main/java/com/example/demo/repository/refreshtoken/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.example.demo.repository;
+package com.example.demo.repository.refreshtoken;
 
 public interface RefreshTokenRepository {
