+++ src/main/java/com/example/demo/service/refreshtoken/RefreshTokenService.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.service.refreshtoken;
 
 public interface RefreshTokenService {
